[frontend] Extract CustomView logic to CustomViewTabsWrapper and useCustomViewRoutes
