+++ opencti-platform/opencti-front/src/private/components/common/stix_domain_objects/StixDomainObjectMain.tsx
@@ -1,9 +1,8 @@
 import { ReactElement, ReactNode } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import StixDomainObjectTabsBox, { type StixDomainObjectTabsBoxTab } from './StixDomainObjectTabsBox';
-import RootCustomView from '@components/custom_views/Root';
-import { useCustomViews } from '@components/custom_views/useCustomViews';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
+import useCustomViewRoutes from '@components/custom_views/useCustomViewRoutes';
 
 interface StixDomainObjectMainProps {
   entityType: string;
@@ -21,7 +20,7 @@ const StixDomainObjectMain = ({
   extraRoutes,
 }: StixDomainObjectMainProps) => {
   const tabs = Object.keys(pages) as StixDomainObjectTabsBoxTab[];
-  const { customViews } = useCustomViews(entityType);
+  const customViewRoutes = useCustomViewRoutes({ entityType });
   return (
     <>
       <StixDomainObjectTabsBox
@@ -61,11 +60,7 @@ const StixDomainObjectMain = ({
         {tabs.includes('history') && (
           <Route path="/history" element={pages.history} />
         )}
-        {
-          customViews.map(({ path, id }) => (
-            <Route key={path} path={path} element={<RootCustomView customViewId={id} />} />
-          ))
-        }
+        {...customViewRoutes}
         {extraRoutes}
         <Route path="*" element={<ErrorNotFound />} />
       </Routes>
